% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_table}
\alias{filter_table}
\title{Apply query to a dataframe}
\usage{
filter_table(data = NULL, filters = NULL)
}
\arguments{
\item{data}{data frame to filter}

\item{filters}{output from queryBuilder sent from
shiny app as \code{input$el_out}
where \code{el} is the htmlWidget element}
}
\description{
Filter a data frame using the output of a queryBuilder.
}
