% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterTable}
\alias{filterTable}
\title{filterTable}
\usage{
filterTable(
  filters = NULL,
  data = NULL,
  date_format = NULL,
  output = c("table", "text")
)
}
\arguments{
\item{filters}{output from queryBuilder htmlWidget sent from shiny app as \code{input$el_out}
where \code{el} is the htmlWidget element}

\item{data}{data frame to filter}

\item{date_format}{optional string specifying the date format used with the datepicker
plugin. See \code{?as.Date()}.}

\item{output}{string return either a filtered data frame (table) or a text representation
of the filter (text)}
}
\description{
filter a data frame using the output of a queryBuilder htmlWidget
}
