% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuilder.R
\name{queryBuilder}
\alias{queryBuilder}
\title{queryBuilder}
\usage{
queryBuilder(
  filters,
  plugins = NULL,
  rules = NULL,
  display_errors = FALSE,
  allow_empty = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{filters}{list of list specifying the available filters in the builder.}

\item{plugins}{character vector of option plugins}

\item{display_errors}{When an error occurs on a rule, display an icon with a tooltip
explaining the error.}

\item{allow_empty}{No error will be thrown is the builder is entirely empty.}
}
\description{
queryBuilder
}
