% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuilder-packer.R, R/queryBuilder.R
\name{queryBuilderInput}
\alias{queryBuilderInput}
\title{queryBuilder}
\usage{
queryBuilderInput(inputId, width = "100\%", rules = NULL, filters)

queryBuilderInput(inputId, width = "100\%", rules = NULL, filters)
}
\arguments{
\item{inputId}{Id of input.}

\item{value}{Initial value.}
}
\description{
Incremental button.

Incremental button.
}
\examples{
library(shiny)

ui <- fluidPage(
  queryBuilderInput("theId", 0)
)

server <- function(input, output) {
  observeEvent(input$theId, {
    print(input$theId)
  })
}

if (interactive()) {
  shinyApp(ui, server)
}
library(shiny)

ui <- fluidPage(
  queryBuilderInput("theId", 0)
)

server <- function(input, output) {
  observeEvent(input$theId, {
    print(input$theId)
  })
}

if (interactive()) {
  shinyApp(ui, server)
}
}
