% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recurseFilter}
\alias{recurseFilter}
\title{recurseFilter}
\usage{
recurseFilter(filter = NULL, date_format = NULL)
}
\arguments{
\item{filter}{filters output from queryBuilder htmlWidget}

\item{date_format}{optional date formatting}
}
\value{
string representation of all filters combined
}
\description{
internal recursive function to process filter
}
