% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{resetQueryBuilder}
\alias{resetQueryBuilder}
\title{Reset a queryBuilder widget}
\usage{
resetQueryBuilder(
  inputId,
  label = NULL,
  icon = NULL,
  width = NULL,
  queryBuilderId = NULL,
  ...
)
}
\arguments{
\item{queryBuilderId}{string. Id of the queryBuilder widget you want to reset}
}
\description{
This function is a simple wrapper around a \code{shiny::actionButton}
with an \code{onclick} attribute that will reset a queryBuilder widget.
Inputs are the same as a regular \code{actionbutton} with an additional
\code{queryBuilderId} input that should correspond to the id of the
queryBuilder widget you want to reset
}
