% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuilder.R
\name{updateQueryBuilder}
\alias{updateQueryBuilder}
\title{updateQueryBuilder}
\usage{
updateQueryBuilder(
  inputId,
  setFilters = NULL,
  addFilters = NULL,
  destroy = FALSE,
  reset = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{setFilters}{}
}
\description{
Update a queryBuilder with available methods.
}
